Doc strings and move ChunkSet to own file
